src/main/java/ru/j4j/finder/PhoneDictionary.java
package ru.j4j.finder;

import ru.j4j.finder.model.Person;

import java.util.ArrayList;
import java.util.function.Predicate;

public class PhoneDictionary {
    private final ArrayList<Person> persons = new ArrayList<>();

    public void add(Person person) {
        this.persons.add(person);
    }

    public ArrayList<Person> find(String key) {
        ArrayList<Person> result = new ArrayList<>();
        Predicate<Person> combine = person -> person.getName().contains(key)
                || person.getSurName().contains(key)
                || person.getPhone().contains(key)
                || person.getAddress().contains(key);
        for (Person person : persons) {
            if (combine.test(person)) {
                result.add(person);
            }
        }
        return result;
    }
}
